millionx-phase2/weather-fetcher/requirements.txt
# Weather Fetcher Requirements

requests==2.31.0
kafka-python==2.0.2
python-dotenv==1.0.0
